fan: correct curve for Linux 5.8 threadripper
In kernel 5.4 the threadripper CPU temperature used by this driver was
Tdie.  In 5.8 it's Tctl, which is 27 Cel higher due to a fixed offset.
The curve was specified for the lower temperature, causing fans to run
at 80% when idling under 5.8.  Adjust the curve by 27 Cel at all
points.

See: https://github.com/pop-os/system76-power/issues/200
Signed-off-by: Peter A. Bigot <pab@pabigot.com>